% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atf3_chr1}
\alias{atf3_chr1}
\title{Atf3 binding regions.}
\format{A GRanges object}
\usage{
data(atf3_chr1)
}
\description{
Binding regions for Atf3.
Only includes regions in chr1 to keep
the example data small.
}
\keyword{datasets}
