% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{averageByRegion}
\alias{averageByRegion}
\title{modification of BSAggregate to just return mean per region
averageByRegion(BSDT = BSDT, regionsGRL, jCommand = MIRA:::buildJ(cols = "methylProp", "mean"))}
\usage{
averageByRegion(loadingMat, coordinateDT, GRList, PCsToAnnotate = c("PC1",
  "PC2"), returnQuantile = FALSE)
}
\arguments{
\item{loadingMat}{matrix of loading values. One named column for each PC and
rows are the original dimensions/variables. The $rotation output of prcomp.}

\item{GRList}{a single region set as a GRangesList. If regionSet is a GRanges
object, this can be made with the command: GRangesList(regionSet)}
}
\value{
a data.table with region coordinates and average loading 
values for each region. Has columns chr, start, end, and a column for each
PC in PCsToAnnotate. Regions are not in order along the rows of the data.table.
}
\description{
modification of BSAggregate to just return mean per region
averageByRegion(BSDT = BSDT, regionsGRL, jCommand = MIRA:::buildJ(cols = "methylProp", "mean"))
}
