% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaATACCoord1}
\alias{brcaATACCoord1}
\title{A GRanges object with coordinates for select 
BRCA ATAC-seq peak regions from chr1.}
\format{
A GRanges object
}
\usage{
data(brcaATACCoord1)
}
\description{
Corresponds to 
the rows of brcaATACData1.
The ATAC-seq data is from breast cancer
patients from The Cancer Genome Atlas
(TCGA-BRCA, Corces et. al, 2018, doi: 10.1126/science.aav1898,
https://atacseq.xenahubs.net/download/brca/brca_peak_Log2Counts_dedup).
Coordinates correspond to the hg38 genome version.
Only select regions on chr1 are included to keep
the example data small.
}
\keyword{datasets}
