% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaATACLoadings1}
\alias{brcaATACLoadings1}
\title{This object contains loadings
for PCA of ATAC-seq data.}
\format{A matrix object}
\usage{
data(brcaATACLoadings1)
}
\description{
From breast cancer
patients from The Cancer Genome Atlas
(TCGA-BRCA, ATAC-seq data from 
Corces et. al, 2018, doi: 10.1126/science.aav1898).
Each row corresponds to one region and
the coordinates for these regions are
in the object brcaATACCoord1, (data("brcaATACCoord1"),
hg38 genome). 
Only regions on chr1 are included to keep
the example data small.
}
\keyword{datasets}
