% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaCoord1}
\alias{brcaCoord1}
\title{A data.table, data.frame object with coordinates for cytosines
from chr1 included in the PCA.}
\format{A data.table, data.frame object}
\usage{
data(brcaCoord1)
}
\description{
Corresponds to 
the rows of brcaLoadings1.
DNA methlyation data is from breast cancer
patients from The Cancer Genome Atlas
(TCGA-BRCA, https://portal.gdc.cancer.gov/).
Only cytosines on chr1 are included to keep
the example data small.
}
\keyword{datasets}
