% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaLoadings1}
\alias{brcaLoadings1}
\title{A matrix with loadings}
\format{A matrix object}
\usage{
data(brcaLoadings1)
}
\description{
This object contains loadings
for PCA of DNA methylation data.
DNA methlyation data is Illumina 450k 
microarray data from breast cancer
patients from The Cancer Genome Atlas
(TCGA-BRCA, https://portal.gdc.cancer.gov/).
Each row corresponds to one cytosine and
the coordinates for these cytosines are
in the object brcaCoords1, (data("brcaCoords1"),
hg38 genome). 
Only cytosines on chr1 are included to keep
the example data small.
}
\keyword{datasets}
