% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaMCoord1}
\alias{brcaMCoord1}
\title{A GRanges object with genomic coordinates for cytosines
from chr1 for the package's built-in DNA methylation data}
\format{A GRanges object}
\usage{
data(brcaMCoord1)
}
\description{
Corresponds to 
the rows of brcaMethylData1.
DNA methlyation data is Illumina 450k 
microarray data from breast cancer
patients from The Cancer Genome Atlas
(TCGA-BRCA, https://portal.gdc.cancer.gov/).
Coordinates correspond to the hg38 genome version.
Only selected cytosines on chr1 are included to keep
the example data small.
}
\keyword{datasets}
