% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaMethylData1}
\alias{brcaMethylData1}
\title{A matrix with DNA methylation levels 
from some CpGs on chromosome 1}
\format{A matrix object}
\usage{
data(brcaMethylData1)
}
\description{
This object contains methylation levels (0 to 1)
for select cytosines in chromosome 1 
for TCGA breast cancer patients from
a DNA methylation microarray (Illumina 450k microarray).
Each row corresponds to one cytosine and
the coordinates for these cytosines are
in the object brcaMCoord1, (data("brcaMCoord1"), hg38 genome). 
Only select cytosines on chr1 are included to keep
the example data small. Columns are patients,
with TCGA patient identifiers as column names.
6004 CpGs and 300 patients are included.
DNA methlyation data is Illumina 450k 
microarray data from breast cancer
patients from The Cancer Genome Atlas
(TCGA-BRCA, https://portal.gdc.cancer.gov/).
}
\keyword{datasets}
