% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaPCScores657}
\alias{brcaPCScores657}
\title{A data.frame with principal component scores for 
PCs 1-4 for 657 breast cancer patients as well
as a column with estrogen receptor status.}
\format{
A data.frame object
}
\usage{
data(brcaPCScores657)
}
\description{
This object contains PC scores for 657 patients
for PCs 1-4. Columns are PCs as well
as a column with estrogen receptor status. Rows are patients,
with TCGA patient identifiers as row names. Patients
were selected from all BRCA patients in TCGA based on having complete
metadata information for estrogen receptor status 
and progesterone receptor status as well as 
having 450k microarray data. 
PCA was done on the Illumina 450k
DNA methlyation  
microarray data (TCGA-BRCA, https://portal.gdc.cancer.gov/),
hg38 genome.
}
\keyword{datasets}
