% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{BSBinAggregate}
\alias{BSBinAggregate}
\title{Produced originally for binning Ewing RRBS data across various region sets}
\usage{
BSBinAggregate(BSDT, rangeDT, binCount, minReads = 500,
  byRegionGroup = TRUE, splitFactor = NULL, PCsToAnnotate = PCsToAnnotate)
}
\arguments{
\item{rangeDT}{A data table with the sets of regions to be binned, 
with columns named start, end}

\item{binCount}{Number of bins across the region}

\item{minReads}{Filter out bins with fewer than X reads before returning.}

\item{byRegionGroup}{Pass along to binCount (see ?binCount)}
}
\description{
Produced originally for binning Ewing RRBS data across various region sets
}
