% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COCOA.R
\docType{package}
\name{COCOA}
\alias{COCOA}
\alias{COCOA-package}
\title{Coordinate Covariation Analysis (COCOA)}
\description{
COCOA is a method for understanding epigenetic variation among samples.
COCOA can be used with epigenetic data that includes 
genomic coordinates and an epigenetic signal,
such as DNA methylation and chromatin accessibility 
data. 
To describe the method on a high level, COCOA quantifies 
inter-sample variation with either a supervised or unsupervised 
technique then uses a database of "region sets" to 
annotate the variation among samples. A region set is a set of 
genomic regions that share a biological annotation, 
for instance transcription factor (TF) binding regions, 
histone modification regions, or open chromatin regions. 
COCOA can identify region sets that are associated with
epigenetic variation between samples and
increase understanding of variation in your data.
}
\references{
\url{http://github.com/databio}
}
\author{
John Lawson

Nathan Sheffield
}
