% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{cpgOLMetrics}
\alias{cpgOLMetrics}
\title{Instead of averaging within regions first as BSAggregate does,
this function does a simple average and standard deviation
 for all CpGs that overlap
with regions of a region set, also does average and 
standard deviation for non overlapping CpGs. Created to 
get metrics of loading values for each PC.}
\usage{
cpgOLMetrics(dataDT, regionGR, metrics = c("mean", "sd"),
  columnMeans = NULL, alsoNonOLMet = TRUE)
}
\arguments{
\item{dataDT}{a data.table with chr, start, end columns as well
as columns to get metrics of eg (PC1, PC2). All columns
except chr, start, and end will be considered 
columns to get the metrics from so no unnecessary columns should be
included.}

\item{alsoNonOLMet}{also include same metrics
for non overlapping CpGs}
}
\description{
Faster if given total average for each column of interest
}
