% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esr1_chr1}
\alias{esr1_chr1}
\title{Estrogen receptor alpha binding regions.}
\format{A GRanges object}
\usage{
data(esr1_chr1)
}
\description{
Binding regions for estrogen receptor alpha (ESR1).
Only includes regions in chr1 to keep
the example data small.
}
\keyword{datasets}
