% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{featuresAlongPC}
\alias{featuresAlongPC}
\title{Look at features (eg, DNA methylation values) in regions of 
interest across samples, 
with samples ordered according to PC of interest. 
The ComplexHeatmap package
is used and additional parameters for the ComplexHeatmap::Heatmap function
may be passed to this function to modify the heatmap.}
\usage{
featuresAlongPC(methylData, mCoord, regionSet, pcScores, orderByPC = "PC1",
  cluster_columns = FALSE, cluster_rows = FALSE, name = "Feature Value",
  ...)
}
\arguments{
\item{methylData}{DNA methylation levels (0 to 1) in matrix or data.frame. 
Rows are cytosines. Columns are samples.}

\item{mCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{regionSet}{A genomic ranges object with regions corresponding
to the same biological annotation.}

\item{pcScores}{The principal component scores for the samples 
(ie transformed methylation data). Must have subject_ID as row names,
These same subject_IDs must be column names of methylData}

\item{orderByPC}{PC to order samples by (order rows of heatmap by PC score, 
from high to low score)}

\item{cluster_columns}{boolean, whether to cluster columns (the features,
eg DNA methylation values for each CpG).}

\item{cluster_rows}{boolean, whether rows should be clustered. 
This should be kept as FALSE to keep the correct ranking of 
samples/observations according to their PC score.}

\item{name}{character object, legend title}

\item{...}{optional parameters for ComplexHeatmap::Heatmap()}
}
\value{
A heatmap of feature values (eg DNA methylation levels) 
in regions of interest (regionSet).
Each row is a patient/sample and each column is an individual feature. 
Rows are ordered by PC score (orderByPC), high scores at top and low at 
the bottom.
}
\description{
Look at features (eg, DNA methylation values) in regions of 
interest across samples, 
with samples ordered according to PC of interest. 
The ComplexHeatmap package
is used and additional parameters for the ComplexHeatmap::Heatmap function
may be passed to this function to modify the heatmap.
}
\examples{
data("brcaMethylData1")
data("brcaCoord1")
data("esr1_chr1")
data("brcaPCScores")
featuresAlongPC(methylData=brcaMethylData1,
                mCoord=brcaCoord1,
                regionSet=esr1_chr1,
                pcScores=brcaPCScores,
                orderByPC="PC1", cluster_columns=TRUE)

}
