% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{fitGammaNullDist}
\alias{fitGammaNullDist}
\title{signalCoord=brcaMCoord1, 
                                 GRList=GRangesList(esr1_chr1), 
                                 signalCol=c("PC1", "PC2"), 
                                 scoringMetric="regionMean")}
\usage{
fitGammaNullDist(nullDistDF, method = "mme", force = FALSE)
}
\description{
signalCoord=brcaMCoord1, 
                                 GRList=GRangesList(esr1_chr1), 
                                 signalCol=c("PC1", "PC2"), 
                                 scoringMetric="regionMean")
}
