% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gata3_chr1}
\alias{gata3_chr1}
\title{Gata3 binding regions.}
\format{A GRanges object}
\usage{
data(gata3_chr1)
}
\description{
Binding regions for gata3.
Only includes regions in chr1 to keep
the example data small.
}
\keyword{datasets}
