% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{getGammaPVal}
\alias{getGammaPVal}
\title{Get p value after fitting a gamma distribution to the null distribution}
\usage{
getGammaPVal(scores, nullDistList, method = "mme",
  realScoreInDist = FALSE, force = FALSE)
}
\arguments{
\item{scores}{a data.frame. Has same columns as nullDistDF. One row per
region set (should be in same order as nullDistDF) The scores
that will be used to get p values.}

\item{nullDistList}{list of a data.frame. Each list item 
has null distributions for a single 
region set. Each column corresponds to a null distribution for that 
region set for a given variable/sample attribute.}

\item{realScoreInDist}{logical. Should the actual score (from 
test with no permutations) be included in the null distribution 
when fitting the gamma distribution}
}
\value{
Returns a data.frame with p values, one column for each col in
scores and nullDistDF
}
\description{
Get p value after fitting a gamma distribution to the null distribution
}
