% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoa.R
\name{getLoadingProfile}
\alias{getLoadingProfile}
\title{Create a "meta-region" loading profile}
\usage{
getLoadingProfile(loadingMat, signalCoord, GRList,
  PCsToAnnotate = c("PC1", "PC2"), binNum = 25, verbose = TRUE)
}
\arguments{
\item{loadingMat}{matrix of loadings (the coefficients of 
the linear combination that defines each PC). One named column for each PC.
One row for each original dimension/variable (should be same order 
as original data/signalCoord). The x$rotation output of prcomp().}

\item{signalCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{GRList}{GRangesList object. Each list item is 
a distinct region set (regions that correspond to 
the same biological annotation). Must be from the same reference genome
as the coordinates for the actual data/samples (signalCoord).}

\item{PCsToAnnotate}{A character vector with principal components to 
include. eg c("PC1", "PC2")}

\item{binNum}{Number of bins to split the regions into when
making the aggregate loading profile. More bins will
give a higher resolution but perhaps more noisy profile.}

\item{verbose}{A "logical" object. Whether progress 
of the function should be shown, one
bar indicates the region set is completed.}
}
\value{
A list of data.tables each data.table for
a separate region set. The data table has the binned loading profile,
one row per bin.
}
\description{
This loading profile can show enrichment 
of cytosines with high loading values in region set but not in
surrounding genome, suggesting that variation is linked specifically
to that region set.
}
\details{
All regions in a given region set 
are combined into a single aggregate profile. Regions should be
expanded on each side to include a wider area of the genome around
the regions of interest. To make the profile, first we take 
the absolute value of the loadings. Then each region is
split into `binNum` bins. All loadings in each bin are 
averaged to get one value per bin. Finally, corresponding bins from
the different regions are averaged (eg all bin1's averaged with each other, 
all bin2's averaged with each other, etc.) to get a single "meta-region"
loading profile. A peak in the middle of this profile suggests
that variability is specific to the region set of interest and is 
not a product of the surrounding genome. A region set can still be
significant even if it does not have a peak. For example, some
}
\examples{
data("brcaMCoord1")
data("brcaLoadings1")
data("esr1_chr1")
GRList = GRangesList(esr1_chr1)
getLoadingProfile(loadingMat=brcaLoadings1, 
                    signalCoord=brcaMCoord1, 
                    GRList=GRList, 
                    PCsToAnnotate=c("PC1", "PC2"), 
                    binNum=25)
}
