% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{loadingProfileSNR}
\alias{loadingProfileSNR}
\title{different scoring metrics
remniscent of LOLA: 
support (number of regions, for us regions that have at least 1 cytosine and can be scored)
mean loading value (or could do ratio of peak to surrounding regions)
signal to noise ratio, how big is peak compared to noise of surrounding area
with SNR, even a small peak could have a high SNR}
\usage{
loadingProfileSNR()
}
\description{
different scoring metrics
remniscent of LOLA: 
support (number of regions, for us regions that have at least 1 cytosine and can be scored)
mean loading value (or could do ratio of peak to surrounding regions)
signal to noise ratio, how big is peak compared to noise of surrounding area
with SNR, even a small peak could have a high SNR
}
