% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{methylAlongPC}
\alias{methylAlongPC}
\title{raw methylation at top enriched regions for a single region set and single PC.
Patients are ordered by PC score for given PC
#' TODO: deal with bug when nrow(highVariable) = 1}
\usage{
methylAlongPC(loadingMat, loadingThreshold, pcScores, coordinateDT, methylData,
  GRList, orderByPC, topXRegions = 50)
}
\arguments{
\item{loadingMat}{matrix of loadings (the coefficients of 
the linear combination that defines each PC). One named column for each PC.
One row for each original dimension/variable (should be same order 
as original data/mCoord). The x$rotation output of prcomp().}

\item{loadingThreshold}{Only select regions with average 
loading at least this high. Based on loading values from orderByPC.}

\item{pcScores}{The principal component scores for the samples
 (ie transformed methylation data). The $x output of prcomp() but must 
have subject_ID as row names.
These same subject_IDs must be column names of methylData.}

\item{methylData}{DNA methylation levels (0 to 1) in matrix or data.frame. 
Rows are cytosines. Columns are samples.}

\item{GRList}{GRangesList object. Each list item is 
a distinct region set (regions that correspond to 
the same biological annotation).}

\item{orderByPC}{PC to order samples by (order rows of heatmap by PC score, 
from high to low score)
REMOVE: rsInd}

\item{topXRegions}{max number of regions to plot, avoids excessively large 
plots which can be hard to load. Number of regions on plot will be less
than or equal to topXRegions (less than if there are not that many regions
total). 50 is arbitrary}
}
\description{
raw methylation at top enriched regions for a single region set and single PC.
Patients are ordered by PC score for given PC
#' TODO: deal with bug when nrow(highVariable) = 1
}
