% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrf1_chr1}
\alias{nrf1_chr1}
\title{Nrf1 binding regions.}
\format{A GRanges object}
\usage{
data(nrf1_chr1)
}
\description{
Binding regions for Nrf1.
Only includes regions in chr1 to keep
the example data small.
}
\keyword{datasets}
