% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{pcFromSubset}
\alias{pcFromSubset}
\title{Function to see how representative the cytosines in a region set are:
can cytosines from a region set reproduce PC ordering and how high is 
the correlation of original PCs with ordering from subset of cytosines?}
\usage{
pcFromSubset(regionSet, mPCA, methylData, mCoord, PCofInterest = "PC1",
  returnCor = FALSE)
}
\arguments{
\item{regionSet}{The region set to subset by. Only loading values for
cytosines in these regions will be used to make a PC score.}

\item{mPCA}{The pca data. Output from prcomp() of DNA methylation data. 
Should have 
mPCA$rotation with PCA loadings, mPCA$x with PC scores and 
mPCA$center with values used to center DNA methylation.}

\item{methylData}{DNA methylation levels in matrix or data.frame. 
Rows are cytosines. Columns are 
samples.}

\item{mCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{returnCor}{Option to return correlation between these scores and
original PC scores.}

\item{pc}{A character object with PCs of interest (eg "PC1"). Loadings from
this PC will be used to generate new PC scores based on only loadings
that overlap with regionSet. Can be multiple PCs.}
}
\value{
a score for each patient from only loading values for cytosines in 
regionSet. If returnCor = TRUE, returns a correlation score for scores from
each "pc" with scores from a subset of loading values for that PC.
}
\description{
Function to see how representative the cytosines in a region set are:
can cytosines from a region set reproduce PC ordering and how high is 
the correlation of original PCs with ordering from subset of cytosines?
}
