% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\docType{package}
\name{PCRSA}
\alias{PCRSA}
\alias{PCRSA-package}
\title{Principal Component Region Set Analysis (PCRSA)}
\description{
A method for annotating principal components of DNA methylation data 
with region sets. A region set is a set of genomic regions that share 
a biological annotation. PCRSA can identify biologically meaningful 
sources of variation along the PCs, increasing interpretability of the PCs 
and understanding of variation in the data.
}
\references{
\url{http://github.com/databio}
}
\author{
John Lawson

Nathan Sheffield
}
