% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{percentCOverlap}
\alias{percentCOverlap}
\title{Determine what percent of cytosines in region sets overlap.
Slightly different than percent overlap of region sets.}
\usage{
percentCOverlap(coordGR, GRList)
}
\arguments{
\item{coordGR}{GenomicRanges object with coordinates for cytosines 
included in the PCA.}

\item{GRList}{GRangesList object. Each list item is 
a distinct region set (regions that correspond to 
the same biological annotation).}
}
\value{
Total number of cytosines from methylation data that overlapped with
each region set. Also the percent of cytosines that overlapped with
other region set.
}
\description{
Determine what percent of cytosines in region sets overlap.
Slightly different than percent overlap of region sets.
}
