\name{print}
\alias{print.ObjCclass}
\alias{print.ObjCid}
\alias{print.ObjCsel}
\title{
  Print methods for ObjC classes
}
\description{
  The following methods print a representation of reference objects
  internal to the Cocoa package in a human-readable form.
}
\usage{
print.ObjCclass(x, ...)
print.ObjCid(x, ...)
print.ObjCsel(x, ...)
}
\arguments{
  \item{x}{object to print}
  \item{...}{additional parameters - currently ignored}
}
%\value{
%}
%\details{
%}
%\seealso{
%}
\keyword{interface}
