% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{regionPCHeatmap}
\alias{regionPCHeatmap}
\title{Heatmap of average loading score for individual regions across PCs.
For a single region set.}
\usage{
regionPCHeatmap()
}
\description{
Heatmap of average loading score for individual regions across PCs.
For a single region set.
}
