% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{regionQuantileByPC}
\alias{regionQuantileByPC}
\title{plot individual region scores/percentiles across PCs for a single region set
One plot for each region set}
\usage{
regionQuantileByPC(loadingMat, coordinateDT, GRList, rsNames,
  PCsToAnnotate = paste0("PC", 1:5))
}
\description{
plot individual region scores/percentiles across PCs for a single region set
One plot for each region set
}
