% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{rsEnrichHeatmap}
\alias{rsEnrichHeatmap}
\title{Heatmap of enrichment scores across PCs
A visualization of the enrichment data.table.}
\usage{
rsEnrichHeatmap(rsEnrichment, PCsToAnnotate = paste0("PC", 1:5),
  orderByPC = "PC1", rsNameCol = "rsNames", topX = 20)
}
\arguments{
\item{PCsToAnnotate}{A character vector with principal components to 
include in the plot, eg c("PC1", "PC2")}

\item{orderByPC}{PC to order by (decreasing order) in heatmap}
}
\description{
Heatmap of enrichment scores across PCs
A visualization of the enrichment data.table.
}
