% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{rsRankingIndex}
\alias{rsRankingIndex}
\title{A convenience function to easily get indices for
the top region sets that were enriched for each PC. 
For each PC, get index of original region sets but ordered by rsEnrichment
ranking for each PC. First number in a given column will be 
original index of the region set ranked first for that PC. Second row for a
column will be the original index of the region set that ranked second
for that PC, etc.}
\usage{
rsRankingIndex(rsEnrichment, PCsToAnnotate)
}
\arguments{
\item{rsEnrichment}{Should be in the same order as GRList (the list of 
region sets used to create it.)}

\item{PCsToAnnotate}{PCs in rsEnrichment for which you want
the indices of the original region sets}
}
\value{
A data.table with columns PCsToAnnotate. Each column has been 
ranked by enrichment score for region sets for that PC.
Original indices for region sets that were used to create rsEnrichment
are given.
}
\description{
Use this function when you want to look at top region sets to make it 
easier to get the original indices to select them from a list of region sets.
}
