% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rsScores}
\alias{rsScores}
\title{Example COCOA Results (made up)}
\format{A data.frame object}
\usage{
data(rsScores)
}
\description{
A data.frame with example COCOA results.
5 region sets with names given by rsScores$rsName.
Each region set has a score for each PC. Scores
for real region sets would normally be orders of 
magnitude smaller.
}
\keyword{datasets}
