% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{rsWilcox}
\alias{rsWilcox}
\title{Wilcoxon rank sum test for a region set}
\usage{
rsWilcox(dataDT, regionGR, ...)
}
\arguments{
\item{dataDT}{a data.table with chr, start, end columns as well
as columns to get metrics of eg (PC1, PC2). All columns
except chr, start, and end will be considered 
columns to get the metrics from so no unnecessary columns should be
included.}

\item{regionGR}{Region set, GenomicRanges object}

\item{...}{Additional parameters of wilcox.test function. See ?wilcox.test.
For instance specify alternative hypothesis: alternative = "greater".}
}
\value{
A vector with a p value for each column other than chr, start or end.
}
\description{
Wilcoxon rank sum test for a region set
}
