% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{runCOCOAPerm}
\alias{runCOCOAPerm}
\title{permutation test by shuffling sample labels}
\usage{
runCOCOAPerm(genomicSignal, signalCoord, GRList, realRSScores,
  sampleLabels, signalCol = c("PC1", "PC2"),
  signalCoordType = "default", scoringMetric = "default",
  variationMetric = "cor", nPerm = 300, useSimpleCache = TRUE,
  cacheDir = getwd(), dataID = "tmp", correctionMethod = "BH", ...)
}
\arguments{
\item{sampleLabels}{data.frame/matrix. Sample labels/values that 
you are running COCOA to find region sets associated with. These 
values will be shuffled for the permutation test. Rows are samples.
Each column is a sample label.}

\item{variationMetric}{character. Either "cor" (correlation), "pcor" (partial
correlation), or "cov" (covariation)}

\item{nPerm}{numeric. The number of permutations to do.}

\item{useSimpleCache}{logical.}

\item{cacheDir}{character.}

\item{dataID}{character. A unique identifier for this dataset 
(for saving results)}

\item{correctionMethod}{character. P value correction method. Default
is "BH" for Benjamini and Hochberg false discovery rate. For acceptable 
arguments and more info see ?stats::p.adjust() (method parameter)}

\item{...}{character. Optional additional arguments for simpleCache}

\item{colsToAnnotate}{character. The column names of `sampleLabels` that
you want to test.}

\item{resultType}{character. "pval" or "zscore"}
}
\value{
Returns a list where each item is a data.frame of COCOA results 
from a separate permutation
}
\description{
For reproducibility, set seed with 'set.seed()' function before running.
}
