% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{signalAlongPC}
\alias{signalAlongPC}
\title{Visualize how genomic signal in a region set changes along principal component axis}
\usage{
signalAlongPC(genomicSignal, signalCoord, regionSet, pcScores,
  orderByPC = "PC1", cluster_columns = FALSE, cluster_rows = FALSE,
  row_title = "Sample", column_title = "Genomic Signal",
  column_title_side = "bottom", name = "Genomic Signal Value", ...)
}
\arguments{
\item{genomicSignal}{DNA methylation levels (0 to 1) in matrix or data.frame. 
Rows are cytosines. Columns are samples.}

\item{signalCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{regionSet}{A genomic ranges object with regions corresponding
to the same biological annotation. Must be from the same reference genome
as the coordinates for the actual data (signalCoord).}

\item{pcScores}{A matrix. The principal component scores for the samples 
(ie transformed methylation data). Must have subject_ID as row names,
These same subject_IDs must be column names of genomicSignal}

\item{orderByPC}{PC to order samples by (order rows of heatmap by PC score, 
from high to low score)}

\item{cluster_columns}{"logical" object, whether to cluster columns (the genomic signal,
eg DNA methylation values for each CpG).}

\item{cluster_rows}{"logical" object, whether rows should be clustered. 
This should be kept as FALSE to keep the correct ranking of 
samples/observations according to their PC score.}

\item{row_title}{character object, row title}

\item{column_title}{character object, column title}

\item{column_title_side}{character object, where to put the column title:
"top" or "bottom"}

\item{name}{character object, legend title}

\item{...}{optional parameters for ComplexHeatmap::Heatmap() (eg change
heatmap colors with "col" parameter)}
}
\value{
A heatmap of genomic signal values (eg DNA methylation levels) 
in regions of interest (regionSet).
Each row is a patient/sample and each column is an individual genomic signal value. 
Rows are ordered by PC score (orderByPC), high scores at top and low at 
the bottom.
}
\description{
Look at genomic signal (eg, DNA methylation values) in regions of 
interest across samples, 
with samples ordered according to score for PC of interest. 
The ComplexHeatmap package
is used and additional parameters for the ComplexHeatmap::Heatmap function
may be passed to this function to modify the heatmap.
}
\examples{
data("brcaMethylData1")
data("brcaMCoord1")
data("esr1_chr1")
data("brcaPCScores")
signalHM <- signalAlongPC(genomicSignal=brcaMethylData1,
                             signalCoord=brcaMCoord1,
                             regionSet=esr1_chr1,
                             pcScores=brcaPCScores,
                             orderByPC="PC1", cluster_columns=TRUE)

}
