% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{aggregateLoadings}
\alias{aggregateLoadings}
\title{Function to aggregate PCA loading weights over a given region set
and then get p value for each PC based on a permutation}
\usage{
aggregateLoadings(loadingMat, coordinateDT, regionSet,
  PCsToAnnotate = c("PC1", "PC2"), permute = FALSE, metric = "meanDiff",
  pcLoadAv = NULL)
}
\arguments{
\item{loadingMat}{}

\item{regionSet}{A genomic ranges object with regions corresponding
to the same biological annotation.
#UPDATE: make sure only aggregating PCsToAnnotate to save time
# permute=TRUE is deprecated and that code chunk is not up to date}

\item{metric}{Scoring metric. Mean difference ("meanDiff") or just a simple
average of the loading values with no normalization ("raw").}

\item{pcLoadAv}{The average absolute loading value for each PC. Will
significantly speed up computation if this is given.}
}
\description{
Function to aggregate PCA loading weights over a given region set
and then get p value for each PC based on a permutation
}
