% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoa.R
\name{aggregateLoadings}
\alias{aggregateLoadings}
\title{Use PCA loadings to score a region set}
\usage{
aggregateLoadings(loadingMat, signalCoord, regionSet,
  PCsToAnnotate = c("PC1", "PC2"), scoringMetric = "regionMean",
  pcLoadAv = NULL, verbose = FALSE)
}
\arguments{
\item{loadingMat}{matrix of loadings (the coefficients of 
the linear combination that defines each PC). One named column for each PC.
One row for each original dimension/variable (should be same order 
as original data/signalCoord). The x$rotation output of prcomp().}

\item{signalCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{regionSet}{A genomic ranges object with regions corresponding
to the same biological annotation. Must be from the same reference genome
as the coordinates for the actual data/samples (signalCoord).}

\item{PCsToAnnotate}{A character vector with principal components to 
include. eg c("PC1", "PC2")}

\item{scoringMetric}{Scoring metric. "regionMean" is a weighted
average of the absolute value of the loadings
with no normalization (recommended). First loadings are
averaged within each region, then all the regions are averaged. With
"regionMean" score, be cautious in interpretation for
 region sets with low number of regions.
Mean difference ("meanDiff") is also supported but is skewed toward 
ranking large region sets highly. Wilcoxon rank sum test ("rankSum")
also is skewed toward ranking large region sets highly and is
significantly slower than the "regionMean" method.}

\item{pcLoadAv}{The average absolute loading value for each PC. Will
significantly speed up computation if this is given.}

\item{verbose}{A "logical" object. Whether progress 
of the function should be shown, one
bar indicates the region set is completed. Useful when using 
aggregateLoadings with 'apply' to do many region sets at a time.}
}
\value{
a data.table with one row and the following 
columns: one column for each item of PCsToAnnotate with name given
by the item which have scores for each PC, cytosine_coverage which
has number of cytosines that overlapped with regionSet,
region_coverage which has number of regions from regionSet
that overlapped any cytosines, total_region_number that has
number of regions in regionSet, mean_region_size that has average
size in base pairs of regions in regionSet, the average is based on
all regions in regionSet and not just ones that overlap.
}
\description{
First, this function identifies which loadings are within the region set. 
Then the loadings are used to score the region set 
according to the `metric` parameter.
}
