% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{averageByRegion}
\alias{averageByRegion}
\title{modification of BSAggregate to just return mean per region}
\usage{
averageByRegion(loadingMat, mCoord, regionSet, PCsToAnnotate = c("PC1",
  "PC2"), returnQuantile = FALSE)
}
\arguments{
\item{loadingMat}{matrix of loadings (the coefficients of 
the linear combination that defines each PC). One named column for each PC.
One row for each original dimension/variable (should be same order 
as original data/mCoord). The x$rotation output of prcomp().}

\item{mCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{regionSet}{A GRanges object with regions corresponding
to the same biological annotation.}

\item{returnQuantile}{Boolean. If FALSE, return region averages. If TRUE,
for each region, return the quantile of that region's average value
based on the distribution of individual cytosine values}
}
\value{
a data.table with region coordinates and average loading 
values for each region. Has columns chr, start, end, and a column for each
PC in PCsToAnnotate. Regions are not in order along the rows of the data.table.
}
\description{
modification of BSAggregate to just return mean per region
}
