% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcaMethylData1}
\alias{brcaMethylData1}
\title{A matrix with DNA methylation levels 
from chromosome 1 for four patients.}
\format{A matrix object}
\usage{
data(brcaMethylData1)
}
\description{
This object contains methylation levels (0 to 1)
for cytosines in chromosome 1 that were covered by
the DNA methylation microarray (Illumina 450k microarray).
Each row corresponds to one cytosine and
the coordinates for these cytosines are
in the object brcaMCoord1, (data("brcaMCoord1"), hg38 genome). 
Only cytosines on chr1 are included to keep
the example data small. Columns are patients,
with TCGA patient identifiers as column names. The patients
were selected based on their PC1 score from
the PCA of DNA methylation on all chromosomes. The
patients with the two highest PC1 scores and 
the two lowest PC1 scores are included 
(see data("brcaPCScores") for the actual scores).
DNA methlyation data is Illumina 450k 
microarray data from breast cancer
patients from The Cancer Genome Atlas
(TCGA-BRCA, https://portal.gdc.cancer.gov/).
}
\keyword{datasets}
