% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{comparePCHeatmap}
\alias{comparePCHeatmap}
\title{create pdf with multiple heatmap plots (number = length(PCsToRankBy)). 
Plot i will be ranked by PCsToRankBy[i]. A wrapper for rsScoreHeatmap}
\usage{
comparePCHeatmap(rsScores, PCsToRankBy = paste0("PC", 1:5),
  PCsToInclude = paste0("PC", 1:10), fileName = NULL, topX = 40)
}
\arguments{
\item{rsScores}{a data.table with scores for each 
region set from main PCRSA function. 
Each row is a region set. Columns are PCs and info on region set overlap
with DNA methylation data. Should be in the same order as GRList (the list of 
region sets used to create it.)}

\item{PCsToRankBy}{PC to order by (decreasing order) in heatmap. One 
heatmap for each PC in PCsToRankBy.}

\item{PCsToInclude}{A character vector with names of PCs that 
should be present in the heatmap.}

\item{fileName}{A character vector. All plots from this function will
be saved to a single pdf. fileName should give the name of that file.
By default, it will be saved in the working directory but filename
can also include the a file path to save the plot in another directory}

\item{topX}{Number of top region sets to include in the heatmap

# see https://github.com/jokergoo/ComplexHeatmap/issues/110}
}
\description{
create pdf with multiple heatmap plots (number = length(PCsToRankBy)). 
Plot i will be ranked by PCsToRankBy[i]. A wrapper for rsScoreHeatmap
}
