% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{comparePCHeatmap}
\alias{comparePCHeatmap}
\title{create pdf with multiple heatmap plots (number = length(PCsToRankBy)). 
Plot i will be ranked by PCsToRankBy[i]. A wrapper for rsScoreHeatmap}
\usage{
comparePCHeatmap(rsScores, PCsToRankBy = paste0("PC", 1:5),
  PCsToInclude = paste0("PC", 1:10), fileName = NULL)
}
\arguments{
\item{rsScores}{a data.table with scores for each 
region set from main PCRSA function. 
Each row is a region set. Columns are PCs and info on region set overlap
with DNA methylation data. Should be in the same order as GRList (the list of 
region sets used to create it.)}

\item{PCsToRankBy}{PC to order by (decreasing order) in heatmap. One 
heatmap for each PC in PCsToRankBy.}

\item{fileName}{# see https://github.com/jokergoo/ComplexHeatmap/issues/110}
}
\description{
create pdf with multiple heatmap plots (number = length(PCsToRankBy)). 
Plot i will be ranked by PCsToRankBy[i]. A wrapper for rsScoreHeatmap
}
