% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{methylAlongPC}
\alias{methylAlongPC}
\title{raw methylation at top enriched regions for a single region set and single PC.
Patients are ordered by PC score for given PC
#' TODO: deal with bug when nrow(highVariable) = 1}
\usage{
methylAlongPC(loadingMat, loadingThreshold, pcScores, coordinateDT, methylData,
  GRList, orderByPC, topXRegions = 50)
}
\arguments{
\item{loadingThreshold}{Only select regions with average 
loading at least this high. Based on loading values from orderByPC.}

\item{orderByPC}{PC to order patients by (order rows of heatmap by PC score)
REMOVE: rsInd}

\item{topXRegions}{max number of regions to plot, avoids excessively large 
plots which can be hard to load. Number of regions on plot will be less
than or equal to topXRegions (less than if there are not that many regions
total). 50 is arbitrary}
}
\description{
raw methylation at top enriched regions for a single region set and single PC.
Patients are ordered by PC score for given PC
#' TODO: deal with bug when nrow(highVariable) = 1
}
