% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{pcEnrichmentProfile}
\alias{pcEnrichmentProfile}
\title{Function to create "meta-region" loading profile that indicates enrichment 
of cytosines with high loading values in region set but not in
surrounding genome. All regions in a given region set 
are combined into a single aggregate profile.}
\usage{
pcEnrichmentProfile(loadingMat, mCoord, GRList, PCsToAnnotate = c("PC1",
  "PC2"), binNum = 25, verbose = TRUE)
}
\arguments{
\item{loadingMat}{matrix of loadings (the coefficients of 
the linear combination that defines each PC). One named column for each PC.
One row for each original dimension/variable (should be same order 
as original data/mCoord). The x$rotation output of prcomp().}

\item{mCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{GRList}{GRangesList object. Each list item is 
a distinct region set (regions that correspond to 
the same biological annotation).}

\item{PCsToAnnotate}{A character vector with principal components to 
include. eg c("PC1", "PC2")}

\item{binNum}{Number of bins to split the regions into when
making the aggregate loading profile. More bins will
give a higher resolution but perhaps more noisy profile.}

\item{verbose}{A boolean. Whether progress of the function should be shown, one
bar for each region set completed.}
}
\value{
A list of data.tables each data.table for
a separate region set. The data table has the binned loading profile,
one row per bin.
}
\description{
Function to create "meta-region" loading profile that indicates enrichment 
of cytosines with high loading values in region set but not in
surrounding genome. All regions in a given region set 
are combined into a single aggregate profile.
}
\examples{
data("brcaCoord1")
data("brcaLoadings1")
data("esr1_chr1")
data("nrf1_chr1")
GRList = GRangesList(esr1_chr1, nrf1_chr1)
pcEnrichmentProfile(loadingMat=brcaLoadings1, mCoord=brcaCoord1, GRList=GRList, PCsToAnnotate=c("PC1", "PC2"), binNum=25)
}
