% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{pcEnrichmentProfile}
\alias{pcEnrichmentProfile}
\title{function to create profile that indicates enrichment 
of cytosines with high loading values in region set but not in
surrounding genome}
\usage{
pcEnrichmentProfile(loadingMat, coordinateDT, GRList, PCsToAnnotate = c("PC1",
  "PC2"), binNum = 25)
}
\description{
A wrapper of BSAggregate that first bins regions and then aggregates
each bin across a set of regions, individually.
}
