% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{pcRegionSetEnrichment}
\alias{pcRegionSetEnrichment}
\title{Wrapper function to do PCA region set enrichment 
analysis for many region sets}
\usage{
pcRegionSetEnrichment(loadingMat, coordinateDT, GRList,
  PCsToAnnotate = c("PC1", "PC2"), permute = TRUE)
}
\value{
data.table of results, one row for each region set. 
One column for each PC in PCsToAnnotate
with average loading score for that PC for a given region set. 
rsName column has region set name.
rsDescription has a description for each region set. 
column has number of cytosines that overlapped with the given region set.
column has number of regions that overlapped with any cytosines.
column has total number of regions.
}
\description{
For parallel processing, region sets are split up between the cores
}
