% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\docType{package}
\name{PCRSA}
\alias{PCRSA}
\alias{PCRSA-package}
\title{Principal Component Region Set Analysis (PCRSA)}
\description{
Purpose of PCRSA/description.
package to annotate PCA components of DNA methylation data
based on region set enrichment
}
\references{
\url{http://github.com/databio}
}
\author{
John Lawson

Nathan Sheffield
}
