% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{regionQuantileByPC}
\alias{regionQuantileByPC}
\title{plot individual region scores/percentiles across PCs for a single region set
One plot for each region set}
\usage{
regionQuantileByPC(loadingMat, mCoord, GRList, rsNames,
  PCsToAnnotate = paste0("PC", 1:5), maxRegionsToPlot = 8000,
  cluster_rows = TRUE)
}
\arguments{
\item{loadingMat}{matrix of loadings (the coefficients of 
the linear combination that defines each PC). One named column for each PC.
One row for each original dimension/variable (should be same order 
as original data/mCoord). The x$rotation output of prcomp().}

\item{mCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{GRList}{GRangesList object. Each list item is 
a distinct region set (regions that correspond to 
the same biological annotation).}

\item{rsNames}{character vector. Names of the region sets in the same
order as GRList. For use as a title for each heatmap.}

\item{PCsToAnnotate}{A character vector with principal components to 
include. eg c("PC1", "PC2")}

\item{maxRegionsToPlot}{how many top regions from region set to include
in heatmap. Including too many may slow down computation and increase memory
use.}

\item{cluster_rows}{Boolean, whether to cluster rows or not (may 
increase computation time significantly for large number of rows)}
}
\description{
plot individual region scores/percentiles across PCs for a single region set
One plot for each region set
}
