% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{rsMethylHeatmap}
\alias{rsMethylHeatmap}
\title{Look at methylation in regions of interest across samples, 
with samples ordered according to PC of interest}
\usage{
rsMethylHeatmap(methylData, coordGR, regionSet, pcScores, orderByPC = "PC1",
  ...)
}
\arguments{
\item{methylData}{DNA methylation levels (0 to 1) in matrix or data.frame. 
Rows are cytosines. Columns are samples.}

\item{coordGR}{coordinates for methylation loci in methylData. (chromosome, start, end)}

\item{regionSet}{}

\item{pcScores}{The principal component scores for the samples 
(ie transformed methylation data). Must have subject_ID as row names,
These same subject_IDs must be column names of methylData}

\item{orderByPC}{PC to order samples by (order rows of heatmap by PC score, 
from high to low score)}
}
\description{
Look at methylation in regions of interest across samples, 
with samples ordered according to PC of interest
}
