% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{rsMethylHeatmap}
\alias{rsMethylHeatmap}
\title{Look at methylation in regions of interest across samples, 
with samples ordered according to PC of interest}
\usage{
rsMethylHeatmap(methylData, mCoord, regionSet, pcScores, orderByPC = "PC1",
  ...)
}
\arguments{
\item{methylData}{DNA methylation levels (0 to 1) in matrix or data.frame. 
Rows are cytosines. Columns are samples.}

\item{mCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{regionSet}{}

\item{pcScores}{The principal component scores for the samples 
(ie transformed methylation data). Must have subject_ID as row names,
These same subject_IDs must be column names of methylData}

\item{orderByPC}{PC to order samples by (order rows of heatmap by PC score, 
from high to low score)}

\item{...}{optional parameters for ComplexHeatmap::Heatmap()}
}
\description{
Look at methylation in regions of interest across samples, 
with samples ordered according to PC of interest
}
