% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRSA.R
\name{rsRankingIndex}
\alias{rsRankingIndex}
\title{A convenience function to easily get indices for
the top region sets that were enriched for each PC. 
For each PC, get index of original region sets but ordered by rsScores
ranking for each PC. First number in a given column will be 
original index of the region set ranked first for that PC. Second row for a
column will be the original index of the region set that ranked second
for that PC, etc.}
\usage{
rsRankingIndex(rsScores, PCsToAnnotate)
}
\arguments{
\item{rsScores}{a data.frame with scores for each 
region set from main PCRSA function. 
Each row is a region set. Columns are PCs and info on region set overlap
with DNA methylation data. Should be in the same order as GRList (the list of 
region sets used to create it.)}

\item{PCsToAnnotate}{PCs in rsScores for which you want
the indices of the original region sets}
}
\value{
A data.frame with columns PCsToAnnotate. Each column has been 
ranked by enrichment score for region sets for that PC.
Original indices for region sets that were used to create rsScores
are given.
}
\description{
Use this function when you want to look at top region sets to make it 
easier to get the original indices to select them from a list of region sets.
}
\examples{
data("rsScores")
rsRankInd = rsRankingIndex(rsScores=rsScores, 
                           PCsToAnnotate=c("PC1", "PC2"))
# region sets sorted by score for PC1
rsScores[rsRankInd$PC1, ]
# region sets sorted by score for PC2
rsScores[rsRankInd$PC2, ]

}
