% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{rsScoreHeatmap}
\alias{rsScoreHeatmap}
\title{Heatmap of enrichment scores across PCs
A visualization of the enrichment data.frame.}
\usage{
rsScoreHeatmap(rsScores, PCsToAnnotate = paste0("PC", 1:5),
  orderByPC = "PC1", rsNameCol = "rsName", topX = 20)
}
\arguments{
\item{rsScores}{a data.table with scores for each 
region set from main PCRSA function. 
Each row is a region set. Columns are PCs and info on region set overlap
with DNA methylation data. Should be in the same order as GRList (the list of 
region sets used to create it.)}

\item{PCsToAnnotate}{A character vector with principal components to 
include. eg c("PC1", "PC2")}

\item{orderByPC}{PC to order by (decreasing order) in heatmap}

\item{rsNameCol}{character. Name of the column in rsScores that has the
names/identifiers for the region sets so this information can be included 
in the plot.}

\item{topX}{Number of top region sets to include in the heatmap}
}
\value{
A heatmap of region set scores across. Each row is a region set,
each column is a PC. The color corresponds to a region set's relative
rank for a given PC out of all tested region sets.
}
\description{
Heatmap of enrichment scores across PCs
A visualization of the enrichment data.frame.
}
