% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoa.R
\name{runCOCOA}
\alias{runCOCOA}
\title{Do COCOA with many region sets}
\usage{
runCOCOA(loadingMat, signalCoord, GRList, PCsToAnnotate = c("PC1",
  "PC2"), scoringMetric = "regionMean", verbose = TRUE)
}
\arguments{
\item{loadingMat}{matrix of loadings (the coefficients of 
the linear combination that defines each PC). One named column for each PC.
One row for each original dimension/variable (should be same order 
as original data/signalCoord). The x$rotation output of prcomp().}

\item{signalCoord}{a GRanges object or data frame with coordinates 
for the cytosines included in the PCA. Coordinates should be in the 
same order as the methylation data and loadings. If a data.frame, 
must have chr and start columns. If end is included, start 
and end should be the same. Start coordinate will be used for calculations.}

\item{GRList}{GRangesList object. Each list item is 
a distinct region set to test (region set: regions that correspond to 
the same biological annotation). Must be from the same reference genome
as the coordinates for the actual data/samples (signalCoord).}

\item{PCsToAnnotate}{A character vector with principal components to 
include. eg c("PC1", "PC2")}

\item{scoringMetric}{Scoring metric. "regionMean" is a weighted
average of the absolute value of the loadings
with no normalization (recommended). First loadings are
averaged within each region, then all the regions are averaged. With
"regionMean" score, be cautious in interpretation for
 region sets with low number of regions.
Mean difference ("meanDiff") is also supported but is skewed toward 
ranking large region sets highly. Wilcoxon rank sum test ("rankSum")
also is skewed toward ranking large region sets highly and is
significantly slower than the "regionMean" method.}

\item{verbose}{A "logical" object. Whether progress 
of the function should be shown, one
bar indicates the region set is completed.}
}
\value{
data.frame of results, one row for each region set. 
One column for each PC in PCsToAnnotate
with score for that PC for a given region set (specific score depends
on "scoringMetric" parameter). 
Rows will be in the same order as region sets in GRList 
and will be named if GRList was named
"cytosine_coverage" column has number of cytosines that 
overlapped with the given region set.
"region_coverage" column has number of regions 
that overlapped with any cytosines.
"total_region_number" column has total number of regions. 
"mean_region_size" has average region size (average of all regions,
not just those that overlap a cytosine).
}
\description{
This function will give each region set a score for each PC
in `PCsToAnnotate` based on
the `scoringMetric` parameter.
}
\examples{
data("brcaMCoord1")
data("brcaLoadings1")
data("esr1_chr1")
rsScores <- runCOCOA(loadingMat=brcaLoadings1, 
                                 signalCoord=brcaMCoord1, 
                                 GRList=GRangesList(esr1_chr1), 
                                 PCsToAnnotate=c("PC1", "PC2"), 
                                 scoringMetric="regionMean")

}
